+++ release.config.js
@@ -1,14 +1,13 @@
-
 module.exports = {
   // eslint-disable-next-line no-template-curly-in-string
   tagFormat: 'v${version}',
-  branch: 'gitlab-ci-runner-docker',
+  branch: 'master',
   prepare: [
     [
       '@semantic-release/changelog',
       {
         changelogFile: 'docs/CHANGELOG.md',
-        changelogTitle: 'df2-skeleton-module changelog documentation ✏️',
+        changelogTitle: 'module-skeleton changelog documentation ✏️',
       },
     ],
     '@semantic-release/npm',
@@ -34,11 +33,6 @@ module.exports = {
         npmPublish: false,
       },
     ],
-    [
-      '@semantic-release/gitlab',
-      {
-        gitlabUrl: 'https://gitlab.ciklum.net',
-      },
-    ],
+    '@semantic-release/github',
   ],
 }