+++ docs/CHANGELOG.md
@@ -1,3 +1,19 @@
+module-skeleton changelog documentation ✏️
+
+# [1.1.0](https://github.com/ciklum-digital/module-skeleton/compare/v1.0.0...v1.1.0) (2019-06-02)
+
+
+### Features
+
+* add jest config ([f220a17](https://github.com/ciklum-digital/module-skeleton/commit/f220a17))
+
+# 1.0.0 (2019-06-02)
+
+
+### Features
+
+* add coveralls integration ([4a2fb3d](https://github.com/ciklum-digital/module-skeleton/commit/4a2fb3d))
+
 df2-skeleton-module changelog documentation ✏️
 
 # [1.1.0](https://gitlab.ciklum.net/df2-components/df2-module-skeleton/compare/v-r1.0.0...v-r1.1.0) (2019-05-21)